// mobile/types/character_types.ts
export type Skill = {
    name: string;
    category: string;
    pureValue: number;
    roleValue: number;
    [key: string]: any;
};